+++ src/ui/image/image-with-shadow/ImageWithShadow.tsx
@@ -19,12 +19,12 @@ export default function ImageWithShadow({
 }: ImageWithShadowProps) {
   return (
     <div
-      aria-hidden='true'
       style={{ height: imageHeight, width: imageWidth }}
       className={cn('relative overflow-hidden', className)}
     >
       <Image {...props} src={src} alt={alt} fill className='object-cover select-none' />
       <div
+        aria-hidden='true'
         className='pointer-events-none absolute inset-0 h-full w-full'
         style={{
           background: 'linear-gradient(180deg, rgba(0, 0, 0, 0) 46.63%, #000 100%)',
